+++ modules/apps/damage.js
@@ -1,11 +1,5 @@
 import { HMPrompt } from './prompt.js';
-import { idx } from '../sys/localize.js';
 import { HMCONST } from '../sys/constants.js';
-
-function getCapList(caps) {
-    const {special} = idx;
-    return Object.fromEntries(caps.map((x) => Object.entries(special)[x]));
-    }
 
 export class DamagePrompt extends HMPrompt {
     static get defaultOptions() {
@@ -18,9 +12,10 @@ export class DamagePrompt extends HMPrompt {
     constructor(dialogData, options) {
         super(dialogData, options);
         const weapon = dialogData.weapons[0];
-        const capList = getCapList(weapon.capabilities);
+        const capList = this.getCapList(weapon.capabilities, dialogData?.callerId);
 
         mergeObject(this.dialogData, {
+            callerId: dialogData?.callerId,
             capList,
             specialMove: 0,
             widx: 0,
@@ -28,22 +23,25 @@ export class DamagePrompt extends HMPrompt {
     }
 
     update(options) {
-        const {weapons, widx} = this.dialogData;
-        this.dialogData.capList = getCapList(weapons[widx].capabilities);
+        const {weapons, widx, callerId} = this.dialogData;
+        this.dialogData.capList = this.getCapList(weapons[widx].capabilities, callerId);
         super.update(options);
     }
 
     get dialogResp() {
         const specialMove = Number(this.dialogData.specialMove);
         const shieldHit = this.dialogData.shieldHit === 'true';
-        const jabbed = specialMove === HMCONST.SPECIAL.JAB;
-
+        const jabbed   = specialMove === HMCONST.SPECIAL.JAB;
+        const backstab = specialMove === HMCONST.SPECIAL.BACKSTAB;
         let formulaType;
         /* eslint 'space-in-parens': 0 */
-        if (!jabbed && !shieldHit) { formulaType = 'standard';  } else
-        if (!jabbed &&  shieldHit) { formulaType = 'shield';    } else
-        if ( jabbed && !shieldHit) { formulaType = 'jab';       } else
-        if ( jabbed &&  shieldHit) { formulaType = 'shieldjab'; }
+
+        if (backstab && !shieldHit) { formulaType = 'bstab';       } else
+        if (backstab &&  shieldHit) { formulaType = 'shieldbstab'; } else
+        if (!jabbed  && !shieldHit) { formulaType = 'standard';    } else
+        if (!jabbed  &&  shieldHit) { formulaType = 'shield';      } else
+        if ( jabbed  && !shieldHit) { formulaType = 'jab';         } else
+        if ( jabbed  &&  shieldHit) { formulaType = 'shieldjab';   }
 
         const dialogResp = {
             widx: this.dialogData.widx,